// src/types.ts
export const FIELDS = [
  'it',
  'engineering',
  'business_management',
  'international_relations',
  'education_languages',
  'natural_sciences',
  'psychology',
] as const;
export type Field = (typeof FIELDS)[number];

export const COUNTRY_PREFERENCES = ['any', 'kz', 'hungary', 'netherlands'] as const;
export type CountryPreference = (typeof COUNTRY_PREFERENCES)[number];